Fix a bug

+++ src/Helper_test.jl
@@ -328,7 +328,7 @@ end
 function RandomSampleDifferentSizeRelativeDensity(B::SparseMatrixCSC, SizeFrom::Int64, SizeUntil::Int64, SizeInterval::Int64, RelativeDensity::Float64, Tests::Int64)
     size = SizeFrom
     while size <= SizeUntil
-        volume = floor(size * (size - 1) * RelativeDensity)
+        volume = Int64(floor(size * (size - 1) * RelativeDensity))
         print_rgb(255,255,128,string("Size = ", size, ": "))
         RandomSampleUntilDensity(B,size,volume,Tests)
         size += SizeInterval
